Fix: Change txnum type to int64 in Clock.FlushAll (master)
Improved the `Clock.FlushAll` function by changing the
`txnum` parameter type from `int` to `int64` to handle
larger transaction numbers.

+++ buffer/Clock.go
@@ -143,7 +143,7 @@ func (c *Clock) Evict() (*Buffer, error) {
 }
 
 // FlushAll implements the EvictionPolicy interface.
-func (c *Clock) FlushAll(txnum int) {
+func (c *Clock) FlushAll(txnum int64) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
